requirements.txt
Flask==3.0.3
Gunicorn==21.2.0
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.11
pydantic==2.9.2
python-dotenv==1.0.1
prometheus-flask-exporter==0.23.1
python-json-logger==2.0.7
